chore: create superclass, add concurrency limit to push
