+++ packages/protocol-identify/README.md
@@ -26,7 +26,16 @@ repo and examine the changes made.
 
 Use the `identify` function to add support for the [Identify protocol](https://github.com/libp2p/specs/blob/master/identify/README.md) to libp2p.
 
-## Example
+This protocol allows network peers to discover the multiaddrs the current node listens on, and the protocols it supports.
+
+A second function, `identifyPush` is also exported to add support for [identify/push](https://github.com/libp2p/specs/blob/master/identify/README.md#identifypush).
+
+This protocol will send updates to all connected peers when the multiaddrs or protocols of the current node change.
+
+> [!TIP]
+> For maximum network compatibility you should configure both protocols
+
+## Example - Enabling identify
 
 ```typescript
 import { createLibp2p } from 'libp2p'
@@ -36,6 +45,20 @@ const node = await createLibp2p({
   // ...other options
   services: {
     identify: identify()
+  }
+})
+```
+
+## Example - Enabling identify push
+
+```typescript
+import { createLibp2p } from 'libp2p'
+import { identifyPush } from '@libp2p/identify'
+
+const node = await createLibp2p({
+  // ...other options
+  services: {
+    identifyPush: identifyPush()
   }
 })
 ```
