+++ packages/protocol-identify/package.json
@@ -58,6 +58,8 @@
     "@libp2p/peer-record": "^7.0.15",
     "@multiformats/multiaddr": "^12.2.1",
     "@multiformats/multiaddr-matcher": "^1.2.0",
+    "it-drain": "^3.0.7",
+    "it-parallel": "^3.0.7",
     "it-protobuf-stream": "^1.1.2",
     "protons-runtime": "^5.4.0",
     "uint8arraylist": "^2.4.8",
@@ -70,7 +72,6 @@
     "@libp2p/peer-id-factory": "^4.1.0",
     "aegir": "^42.2.5",
     "delay": "^6.0.0",
-    "it-drain": "^3.0.5",
     "it-length-prefixed": "^9.0.4",
     "it-pair": "^2.0.6",
     "it-pushable": "^3.2.3",
