+++ packages/protocol-identify/test/push.spec.ts
@@ -7,7 +7,7 @@ import { expect } from 'aegir/chai'
 import { pair } from 'it-pair'
 import { pbStream } from 'it-protobuf-stream'
 import { stubInterface } from 'sinon-ts'
-import { IdentifyPush } from '../src/identify.js'
+import { IdentifyPush } from '../src/identify-push.js'
 import { Identify as IdentifyMessage } from '../src/pb/message.js'
 import { identifyPushStream, type StubbedIdentifyComponents } from './fixtures/index.js'
 import type { Libp2pEvents, PeerStore } from '@libp2p/interface'
@@ -43,7 +43,7 @@ describe('identify (push)', () => {
     await start(identify)
 
     expect(components.registrar.handle.called).to.be.true('identify push not handled')
-    expect(components.registrar.handle.getCall(1).args[0]).to.equal('/ipfs/id/push/1.0.0')
+    expect(components.registrar.handle.getCall(0).args[0]).to.equal('/ipfs/id/push/1.0.0')
   })
 
   it('should be able to push identify updates to another peer', async () => {
@@ -120,7 +120,7 @@ describe('identify (push)', () => {
 
     components.peerStore.patch.reset()
 
-    await identify._handlePush({
+    await identify.handleProtocol({
       stream,
       connection
     })
@@ -149,7 +149,7 @@ describe('identify (push)', () => {
 
     components.peerStore.patch.reset()
 
-    await expect(identify._handlePush({
+    await expect(identify.handleProtocol({
       stream,
       connection
     })).to.eventually.be.undefined()
